pyproject.toml
[project]
name = "django-fullstack-development"
version = "0.1.0"
description = ""
authors = [
    {name = "amanmdest",email = "amanmdest@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14.0"
dependencies = [
    "django",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "mysite.settings"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ipython = "*"
pytest-django = "*"
